green-hash-maroc/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
web3==6.10.0
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
gunicorn==21.2.0
redis==5.0.0
flask-limiter==3.5.0